+++ apps/vs-code-designer/src/app/commands/workflows/openDesigner/openDesignerForAzureResource.ts
@@ -12,8 +12,8 @@ import { getAuthorizationToken } from '../../../utils/codeless/getAuthorizationT
 import type { IAzureConnectorsContext } from '../azureConnectorWizard';
 import { OpenDesignerBase } from './openDesignerBase';
 import type { ServiceClientCredentials } from '@azure/ms-rest-js';
-import type { IDesignerPanelMetadata, IWorkflowFileContent } from '@microsoft/utils-logic-apps';
-import { ExtensionCommand } from '@microsoft/utils-logic-apps';
+import type { IWorkflowFileContent, IDesignerPanelMetadata } from '@microsoft/vscode-extension';
+import { ExtensionCommand } from '@microsoft/vscode-extension';
 import * as path from 'path';
 import * as vscode from 'vscode';