+++ apps/vs-code-designer/src/app/tree/configurationsTree/connectionsTree/ConnectionsTreeItem.ts
@@ -8,9 +8,9 @@ import { getThemedIconPath } from '../../../utils/tree/assets';
 import { getProjectContextValue } from '../../../utils/tree/projectContextValues';
 import type { ConfigurationsTreeItem } from '../ConfigurationsTreeItem';
 import { ConnectionTreeItem } from './ConnectionTreeItem';
-import { ProjectAccess, ProjectResource } from '@microsoft/utils-logic-apps';
 import { AzExtParentTreeItem } from '@microsoft/vscode-azext-utils';
 import type { AzExtTreeItem, ILoadingTreeContext, TreeItemIconPath } from '@microsoft/vscode-azext-utils';
+import { ProjectAccess, ProjectResource } from '@microsoft/vscode-extension';
 
 export class ConnectionsTreeItem extends AzExtParentTreeItem {
   public readonly label: string = localize('Connections', 'Connections');