+++ apps/vs-code-designer/src/app/tree/remoteWorkflowsTree/RemoteWorkflowTreeItem.ts
@@ -13,16 +13,12 @@ import { getProjectContextValue } from '../../utils/tree/projectContextValues';
 import type { RemoteWorkflowsTreeItem } from './RemoteWorkflowsTreeItem';
 import type { StringDictionary } from '@azure/arm-appservice';
 import type { ServiceClientCredentials } from '@azure/ms-rest-js';
-import { ProjectResource, isEmptyString, HTTP_METHODS } from '@microsoft/utils-logic-apps';
-import type {
-  IWorkflowFileContent,
-  Artifacts,
-  Parameter,
-  ICallbackUrlResponse,
-  ServiceProviderConnectionModel,
-} from '@microsoft/utils-logic-apps';
+import { isEmptyString, HTTP_METHODS } from '@microsoft/utils-logic-apps';
 import { AzExtTreeItem, DialogResponses } from '@microsoft/vscode-azext-utils';
 import type { AzExtParentTreeItem, IActionContext, TreeItemIconPath } from '@microsoft/vscode-azext-utils';
+import { ProjectResource } from '@microsoft/vscode-extension';
+import type { ServiceProviderConnectionModel } from '@microsoft/vscode-extension';
+import type { IWorkflowFileContent, Artifacts, Parameter, ICallbackUrlResponse } from '@microsoft/vscode-extension';
 import { ProgressLocation, window } from 'vscode';
 
 export class RemoteWorkflowTreeItem extends AzExtTreeItem {