+++ apps/vs-code-designer/src/app/tree/remoteWorkflowsTree/RemoteWorkflowsTreeItem.ts
@@ -10,10 +10,11 @@ import { getThemedIconPath } from '../../utils/tree/assets';
 import { getProjectContextValue } from '../../utils/tree/projectContextValues';
 import type { SlotTreeItemBase } from '../slotsTree/SlotTreeItemBase';
 import { RemoteWorkflowTreeItem } from './RemoteWorkflowTreeItem';
-import { isNullOrEmpty, ProjectAccess, ProjectResource } from '@microsoft/utils-logic-apps';
-import type { Artifacts, Parameter } from '@microsoft/utils-logic-apps';
+import { isNullOrEmpty } from '@microsoft/utils-logic-apps';
 import { AzExtParentTreeItem } from '@microsoft/vscode-azext-utils';
 import type { AzExtTreeItem, IActionContext, TreeItemIconPath } from '@microsoft/vscode-azext-utils';
+import { ProjectAccess, ProjectResource } from '@microsoft/vscode-extension';
+import type { Artifacts, Parameter } from '@microsoft/vscode-extension';
 
 export class RemoteWorkflowsTreeItem extends AzExtParentTreeItem {
   public readonly label: string = localize('Workflows', 'Workflows');