+++ apps/vs-code-designer/src/app/utils/codeless/apiUtils.ts
@@ -7,16 +7,10 @@ import { localize } from '../../../localize';
 import type { RemoteWorkflowTreeItem } from '../../tree/remoteWorkflowsTree/RemoteWorkflowTreeItem';
 import type { SlotTreeItemBase } from '../../tree/slotsTree/SlotTreeItemBase';
 import { sendAzureRequest } from '../requestUtils';
-import type {
-  Artifacts,
-  IArtifactFile,
-  IConnectionsFileContent,
-  IParametersFileContent,
-  IWorkflowFileContent,
-  Parameter,
-} from '@microsoft/utils-logic-apps';
 import { HTTP_METHODS } from '@microsoft/utils-logic-apps';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
+import type { IConnectionsFileContent } from '@microsoft/vscode-extension';
+import type { IParametersFileContent, Parameter, IWorkflowFileContent, IArtifactFile, Artifacts } from '@microsoft/vscode-extension';
 import * as path from 'path';
 import * as vscode from 'vscode';