+++ apps/vs-code-designer/src/app/utils/codeless/connection.ts
@@ -5,15 +5,11 @@ import { tryGetFunctionProjectRoot } from '../verifyIsProject';
 import { getContainingWorkspace } from '../workspace';
 import { getAuthorizationToken } from './getAuthorizationToken';
 import { getParametersJson } from './parameter';
-import type {
-  ServiceProviderConnectionModel,
-  ConnectionAndSettings,
-  ConnectionReferenceModel,
-  Parameter,
-} from '@microsoft/utils-logic-apps';
 import { HTTP_METHODS } from '@microsoft/utils-logic-apps';
 import { nonNullValue } from '@microsoft/vscode-azext-utils';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
+import type { ServiceProviderConnectionModel, ConnectionAndSettings, ConnectionReferenceModel } from '@microsoft/vscode-extension';
+import type { Parameter } from '@microsoft/vscode-extension';
 import * as fse from 'fs-extra';
 import * as path from 'path';
 import * as requestP from 'request-promise';