+++ apps/vs-code-designer/src/app/utils/fs.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 import { isEmptyString } from '@microsoft/utils-logic-apps';
-import type { pathRelativeFunc } from '@microsoft/utils-logic-apps';
+import type { pathRelativeFunc } from '@microsoft/vscode-extension';
 import * as fse from 'fs-extra';
 import * as path from 'path';