+++ apps/vs-code-designer/src/app/utils/vsCodeConfig/settings.ts
@@ -3,7 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 import { ext } from '../../../extensionVariables';
-import { isString, ProjectLanguage, WorkerRuntime } from '@microsoft/utils-logic-apps';
+import { isString } from '@microsoft/utils-logic-apps';
+import { ProjectLanguage, WorkerRuntime } from '@microsoft/vscode-extension';
 import { Uri, workspace } from 'vscode';
 import type { WorkspaceConfiguration, WorkspaceFolder } from 'vscode';