+++ libs/data-mapper/src/lib/utils/__test__/DataMapUtils.spec.ts
@@ -7,7 +7,7 @@ import {
   targetMockSchema,
 } from '../../__mocks__';
 import type { MapDefinitionEntry } from '../../models';
-import { ifPseudoFunctionKey, functionMock, SchemaType } from '../../models';
+import { functionMock, ifPseudoFunctionKey, SchemaType } from '../../models';
 import type { ConnectionDictionary, ConnectionUnit } from '../../models/Connection';
 import { addParentConnectionForRepeatingElementsNested, convertFromMapDefinition, getSourceValueFromLoop } from '../DataMap.Utils';
 import { convertSchemaToSchemaExtended, flattenSchemaIntoDictionary } from '../Schema.Utils';
@@ -52,35 +52,12 @@ describe('utils/DataMap', () => {
       expect(result['source-/ns0:Root/DirectTranslation/EmployeeName']).toBeTruthy();
     });
 
-    it('creates a simple connection between one source, one function and one target', () => {
+    it('creates a connection between a custom value and target node', () => {
       simpleMap['ns0:Root'] = {
         DirectTranslation: {
           Employee: {
-            Name: 'concat(/ns0:Root/DirectTranslation/EmployeeName)',
-          },
-        },
-      };
-      const result = convertFromMapDefinition(simpleMap, extendedSource, extendedTarget, functionMock);
-      // target lists function as input
-      const targetInput = result['target-/ns0:Root/DirectTranslation/Employee/Name'].inputs['0'][0] as ConnectionUnit;
-      expect(targetInput.reactFlowKey).toContain('Concat');
-
-      // source lists function as output
-      const sourceOutput = result['source-/ns0:Root/DirectTranslation/EmployeeName'].outputs[0];
-      expect(sourceOutput.reactFlowKey).toContain('Concat');
-
-      // function lists source as input
-
-      // function lists target as output
-
-      expect(Object.keys(result).some((key) => key.includes('Concat')));
-    });
-
-    it('creates a conditional connection', () => {
-      simpleMap['ns0:Root'] = {
-        DirectTranslation: {
-          Employee: {
-            Name: '$if(is-greater-than(/ns0:Root/ConditionalMapping/ItemPrice, /ns0:Root/ConditionalMapping/ItemQuantity))',
+            Name: '"Steve"',
+            ID: '10',
           },
         },
       };
@@ -89,7 +66,96 @@ describe('utils/DataMap', () => {
       const resultEntries = Object.entries(result);
       resultEntries.sort();
 
-      expect(resultEntries.length).toEqual(5);
+      expect(resultEntries.length).toEqual(2);
+
+      expect(resultEntries[0][0]).toContain('target-/ns0:Root/DirectTranslation/Employee/ID');
+      expect(resultEntries[0][1]).toBeTruthy();
+      expect(resultEntries[0][1].inputs[0][0]).toEqual('10');
+
+      expect(resultEntries[1][0]).toContain('target-/ns0:Root/DirectTranslation/Employee/Name');
+      expect(resultEntries[1][1]).toBeTruthy();
+      expect(resultEntries[1][1].inputs[0][0]).toEqual('"Steve"');
+    });
+
+    it('creates a simple connection between one source, one function and one target', () => {
+      simpleMap['ns0:Root'] = {
+        DirectTranslation: {
+          Employee: {
+            Name: 'concat(/ns0:Root/DirectTranslation/EmployeeName)',
+          },
+        },
+      };
+
+      const result = convertFromMapDefinition(simpleMap, extendedSource, extendedTarget, functionMock);
+      const resultEntries = Object.entries(result);
+      resultEntries.sort();
+
+      expect(resultEntries.length).toEqual(3);
+
+      const concatId = resultEntries[0][0];
+
+      expect(resultEntries[0][0]).toEqual(concatId);
+      expect(resultEntries[0][1]).toBeTruthy();
+      expect((resultEntries[0][1].inputs[0][0] as ConnectionUnit).reactFlowKey).toEqual('source-/ns0:Root/DirectTranslation/EmployeeName');
+      expect(resultEntries[0][1].outputs[0].reactFlowKey).toEqual('target-/ns0:Root/DirectTranslation/Employee/Name');
+
+      expect(resultEntries[1][0]).toContain('source-/ns0:Root/DirectTranslation/EmployeeName');
+      expect(resultEntries[1][1]).toBeTruthy();
+      expect(resultEntries[1][1].outputs[0].reactFlowKey).toEqual(concatId);
+
+      expect(resultEntries[2][0]).toContain('target-/ns0:Root/DirectTranslation/Employee/Name');
+      expect(resultEntries[2][1]).toBeTruthy();
+      expect((resultEntries[2][1].inputs[0][0] as ConnectionUnit).reactFlowKey).toEqual(concatId);
+    });
+
+    it('creates a connection between a source, a function with custom value and a target', () => {
+      simpleMap['ns0:Root'] = {
+        DirectTranslation: {
+          Employee: {
+            Name: 'concat("Employee Name: ", /ns0:Root/DirectTranslation/EmployeeName, ", Esq")',
+          },
+        },
+      };
+
+      const result = convertFromMapDefinition(simpleMap, extendedSource, extendedTarget, functionMock);
+      const resultEntries = Object.entries(result);
+      resultEntries.sort();
+
+      expect(resultEntries.length).toEqual(3);
+
+      const concatId = resultEntries[0][0];
+
+      expect(resultEntries[0][0]).toEqual(concatId);
+      expect(resultEntries[0][1]).toBeTruthy();
+      expect(resultEntries[0][1].inputs[0][0]).toEqual('"Employee Name: "');
+      expect((resultEntries[0][1].inputs[0][1] as ConnectionUnit).reactFlowKey).toEqual('source-/ns0:Root/DirectTranslation/EmployeeName');
+      expect(resultEntries[0][1].inputs[0][2]).toEqual('", Esq"');
+      expect(resultEntries[0][1].outputs[0].reactFlowKey).toEqual('target-/ns0:Root/DirectTranslation/Employee/Name');
+
+      expect(resultEntries[1][0]).toContain('source-/ns0:Root/DirectTranslation/EmployeeName');
+      expect(resultEntries[1][1]).toBeTruthy();
+      expect(resultEntries[1][1].outputs[0].reactFlowKey).toEqual(concatId);
+
+      expect(resultEntries[2][0]).toContain('target-/ns0:Root/DirectTranslation/Employee/Name');
+      expect(resultEntries[2][1]).toBeTruthy();
+      expect((resultEntries[2][1].inputs[0][0] as ConnectionUnit).reactFlowKey).toEqual(concatId);
+    });
+
+    it.skip('creates a conditional connection', () => {
+      simpleMap['ns0:Root'] = {
+        ConditionalMapping: {
+          '$if(is-greater-than(multiply(/ns0:Root/ConditionalMapping/ItemPrice, /ns0:Root/ConditionalMapping/ItemQuantity), 200))': {
+            ItemDiscount: 'multiply(/ns0:Root/ConditionalMapping/ItemPrice, /ns0:Root/ConditionalMapping/ItemQuantity, 0.05)',
+          },
+        },
+      };
+
+      const result = convertFromMapDefinition(simpleMap, extendedSource, extendedTarget, functionMock);
+      const resultEntries = Object.entries(result);
+      resultEntries.sort();
+
+      // All these expects might be incorrect
+      expect(resultEntries.length).toEqual(6);
 
       expect(resultEntries[0][0]).toContain('IsGreater');
       expect(resultEntries[0][1]).toBeTruthy();
@@ -142,7 +208,7 @@ describe('utils/DataMap', () => {
         ConditionalLooping: {
           CategorizedCatalog: {
             '$for(/ns0:Root/ConditionalLooping/FlatterCatalog/ns0:Product)': {
-              '$if(is-equal(substring(SKU, 1, 2), "11"))': {
+              '$if(is-equal(substring(SKU, 1, 2), "1"))': {
                 PetProduct: {
                   Name: 'Name',
                 },
@@ -155,7 +221,8 @@ describe('utils/DataMap', () => {
       const result = convertFromMapDefinition(simpleMap, extendedSource, extendedTarget, functionMock);
       const resultEntries = Object.entries(result);
       resultEntries.sort();
-      /*
+
+      // All these expects might be incorrect
       expect(resultEntries.length).toEqual(5);
 
       expect(resultEntries[0][0]).toContain('IsGreater');
@@ -172,7 +239,80 @@ describe('utils/DataMap', () => {
 
       expect(resultEntries[4][0]).toEqual('target-/ns0:Root/DirectTranslation/Employee/Name');
       expect(resultEntries[4][1]).toBeTruthy();
-      */
+    });
+
+    it.skip('creates a direct index connection', () => {
+      simpleMap['ns0:Root'] = {
+        LoopingWithIndex: {
+          WeatherSummary: {
+            Day1: {
+              Name: '"Day 1"',
+              Pressure: '/ns0:Root/LoopingWithIndex/WeatherReport[1]/@Pressure',
+            },
+          },
+        },
+      };
+
+      const result = convertFromMapDefinition(simpleMap, extendedSource, extendedTarget, functionMock);
+      const resultEntries = Object.entries(result);
+      resultEntries.sort();
+
+      // All these expects might be incorrect
+      expect(resultEntries.length).toEqual(5);
+
+      expect(resultEntries[0][0]).toContain('IsGreater');
+      expect(resultEntries[0][1]).toBeTruthy();
+
+      expect(resultEntries[1][0]).toContain(ifPseudoFunctionKey);
+      expect(resultEntries[1][1]).toBeTruthy();
+
+      expect(resultEntries[2][0]).toEqual('source-/ns0:Root/ConditionalMapping/ItemPrice');
+      expect(resultEntries[2][1]).toBeTruthy();
+
+      expect(resultEntries[3][0]).toEqual('source-/ns0:Root/ConditionalMapping/ItemQuantity');
+      expect(resultEntries[3][1]).toBeTruthy();
+
+      expect(resultEntries[4][0]).toEqual('target-/ns0:Root/DirectTranslation/Employee/Name');
+      expect(resultEntries[4][1]).toBeTruthy();
+    });
+
+    it.skip('creates a direct index looping connection', () => {
+      simpleMap['ns0:Root'] = {
+        LoopingWithIndex: {
+          WeatherSummary: {
+            '$for(/ns0:Root/LoopingWithIndex/WeatherReport, $a)': {
+              '$if(is-greater-than($a, 2))': {
+                Day: {
+                  Name: 'concat("Day ", $a)',
+                  Pressure: './@Pressure',
+                },
+              },
+            },
+          },
+        },
+      };
+
+      const result = convertFromMapDefinition(simpleMap, extendedSource, extendedTarget, functionMock);
+      const resultEntries = Object.entries(result);
+      resultEntries.sort();
+
+      // All these expects might be incorrect
+      expect(resultEntries.length).toEqual(5);
+
+      expect(resultEntries[0][0]).toContain('IsGreater');
+      expect(resultEntries[0][1]).toBeTruthy();
+
+      expect(resultEntries[1][0]).toContain(ifPseudoFunctionKey);
+      expect(resultEntries[1][1]).toBeTruthy();
+
+      expect(resultEntries[2][0]).toEqual('source-/ns0:Root/ConditionalMapping/ItemPrice');
+      expect(resultEntries[2][1]).toBeTruthy();
+
+      expect(resultEntries[3][0]).toEqual('source-/ns0:Root/ConditionalMapping/ItemQuantity');
+      expect(resultEntries[3][1]).toBeTruthy();
+
+      expect(resultEntries[4][0]).toEqual('target-/ns0:Root/DirectTranslation/Employee/Name');
+      expect(resultEntries[4][1]).toBeTruthy();
     });
   });