+++ libs/utils/src/lib/helpers/hooks.ts
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useRef } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 
 export const useThrottledEffect = (effect: () => void, deps: any[], delay: number) => {
   const timestamp = useRef(Date.now());
@@ -16,4 +16,24 @@ export const useThrottledEffect = (effect: () => void, deps: any[], delay: numbe
     const handler = setTimeout(timeoutFunc, delay - (Date.now() - timestamp.current));
     return () => clearTimeout(handler);
   }, [callback, delay]);
+};
+
+function getWindowDimensions() {
+  const { innerWidth: width, innerHeight: height } = window;
+  return { width, height };
+}
+
+export const useWindowDimensions = () => {
+  const [windowDimensions, setWindowDimensions] = useState(getWindowDimensions());
+
+  useEffect(() => {
+    function handleResize() {
+      setWindowDimensions(getWindowDimensions());
+    }
+
+    window.addEventListener('resize', handleResize);
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return windowDimensions;
 };