+++ CHANGELOG.md
@@ -1,4 +1,24 @@
 # Logic Apps Designer
+### [0.2.40](https://github.com/Azure/LogicAppsUX/compare/v0.2.39...v0.2.40) (2022-12-08)
+
+### [0.2.39](https://github.com/Azure/LogicAppsUX/compare/v0.2.38...v0.2.39) (2022-12-07)
+
+### [0.2.38](https://github.com/Azure/LogicAppsUX/compare/v0.2.37...v0.2.38) (2022-12-07)
+
+### [0.2.37](https://github.com/Azure/LogicAppsUX/compare/v0.2.36...v0.2.37) (2022-12-07)
+
+### [0.2.36](https://github.com/Azure/LogicAppsUX/compare/v0.2.35...v0.2.36) (2022-12-06)
+
+### [0.2.35](https://github.com/Azure/LogicAppsUX/compare/v0.2.34...v0.2.35) (2022-12-05)
+
+### [0.2.34](https://github.com/Azure/LogicAppsUX/compare/v0.2.33...v0.2.34) (2022-12-04)
+
+### [0.2.33](https://github.com/Azure/LogicAppsUX/compare/v0.2.32...v0.2.33) (2022-12-03)
+
+
+### Logic Apps Designer Changes
+
+* Workflow is clamped to remain in view  ([#1349](https://github.com/Azure/LogicAppsUX/issues/1349)) ([ce4adb6](https://github.com/Azure/LogicAppsUX/commit/ce4adb62564a2242811151fe9810aba34b09cb59))
 
 ### [0.2.32](https://github.com/Azure/LogicAppsUX/compare/v0.2.31...v0.2.32) (2022-12-02)